[chore 163304254]Re-implement the create meetup endpoint
